add captcha, fix misc errors

+++ index.php
@@ -1,12 +1,13 @@
 <?php
-	if(file_exists(getcwd().'/oc-install/installer.php') && is_writable(getcwd())){
-		session_start();
+	$DEV = false;
+	if(!$DEV && file_exists(getcwd().'/oc-install/installer.php') && is_writable(getcwd())){
+		if(session_status() === PHP_SESSION_NONE) session_start();
 		session_unset();
 		session_destroy();
 		setcookie('aljksdz7', null, -1, "/");
 		header('Location://'.$_SERVER['SERVER_NAME'].$_SERVER['REQUEST_URI'].'oc-install/installer.php');
 	}
-	if(!file_exists(getcwd().'/oc-install/installer.php')){
+	if(!file_exists(getcwd().'/oc-install/installer.php') && is_dir("oc-install") && !$DEV){
 		 rmdir("oc-install");
 	}
 
@@ -14,7 +15,12 @@
 	require_once(__DIR__ . "/actions/register.php");
 	require_once(__DIR__ . "/actions/publicFunctions.php");
 
-	session_start();
+	if(session_status() === PHP_SESSION_NONE) session_start();
+
+	if(LOGIN_CAPTCHA_ENABLED) {
+		include("plugins/captcha/simple-php-captcha.php");
+		$_SESSION['captcha'] = simple_php_captcha();
+	}
 	
 	if ( (isset($_SESSION['logged_in'])) == "YES" )
 	{
@@ -22,17 +28,17 @@
 	}
 	if (isset($_GET['loggedOut']))
 	{
-	  $loginMessage = '<div class="alert alert-success" style="text-align: center;" ><span>You\'ve successfully been logged out</span></div>';
+	  $loginMessage = "<script> $(function(){ M.toast({html: 'You are now logged out!', classes: 'green accent-4'}); });</script>";
    }
 	if(isset($_SESSION['register_error']))
 	{
-	  $registerError = '<div class="alert alert-danger" style="text-align: center;"><span>'.$_SESSION['register_error'].'</span></div>';
+	  $registerError = "<script> $(function(){ M.toast({html: '${$_SESSION['register_error']}', classes: \'red darken-4\'}); });</script>";
 		unset($_SESSION['register_error']);
 	}
 	if(isset($_SESSION['loginMessageDanger']))
 	{
-	  $loginMessage = '<div class="alert alert-danger" style="text-align: center;"><span>'.$_SESSION['loginMessageDanger'].'</span></div>';
-		unset($_SESSION['loginMessageDanger']);
+	  $loginMessage = sprintf("<script> $(function(){ M.toast({html: '%s', classes: 'red darken-4'}); });</script>", $_SESSION['loginMessageDanger']);
+	  unset($_SESSION['loginMessageDanger']);
 	}
 ?>
 
@@ -45,6 +51,15 @@
    <body>
       <div class="section">
          <h1 class="center-align"> OpenCAD | Login </h1>
+		 <?php 
+		 if ($DEV) {
+		 ?>
+
+		<h2 class="center-align error-text"> <em> !! DEV MODE !! </em> </h2>	
+
+		 <?php 
+		 }
+		 ?>
          <form action="./actions/login.php" method="post">
          <div class="row">
             <div class="input-field col offset-s4 s4">
@@ -56,11 +71,32 @@
             <div class="input-field col offset-s4 s4">
                <input name="password" id="password" type="password" class="validate" required>
                <label for="password">Password</label>
-               <input type="submit" class="btn red darken-4" value="Login">
+            </div>
+		 </div>
+		 <?php if(LOGIN_CAPTCHA_ENABLED) { ?>
+			<div class="row">
+            <div class="input-field col offset-s4 s4">
+			
+			<?php 
+				
+				$captcha_img = sprintf("<img src='%s' style='padding-left: 25vh;'></img>", $_SESSION['captcha']['image_src']);
+				echo $captcha_img;
+
+				?>
+			
+            </div>
+		 </div>
+			<div class="row">
+            <div class="input-field col offset-s4 s4">
+               <input name="captcha" id="captcha" type="text" class="validate" required>
+			   <label for="captcha">Required Captcha (case sensitive)</label>
+
+			   <input type="submit" class="btn red darken-4" value="Login">
                <a class="btn red darken-4" onclick="M.toast({html: 'Contact an admin to reset your password.', classes:'red darken-4'});" >Forgot Password?</a>
                <a class="btn red darken-4 modal-trigger" href="#register">New? Register</a>
             </div>
-         </div>
+		 </div>
+		<?php } ?>
         </form>
       </div>
       <div id="register" class="modal">
@@ -74,7 +110,7 @@
 						<input class="form-control" placeholder="Email" name="email" type="email" required>
 					 </div>
 					 <div class="input-field col s12">
-						<input class="form-control" placeholder="Identifier (Code Number, Unit ID)" name="identifier" type="text" required>
+						<input class="form-control" placeholder="Identifier (1K24, 124, ETC.)" name="identifier" type="text" required>
 					 </div>
 					 <div class="input-field col s12">
 						<input class="form-control" placeholder="Password" name="password" type="password" value="<?php if($testing){echo "password";}?>" required>
@@ -117,8 +153,17 @@
 			if($_GET['register'] == "passwordMatch") {
 				echo "<script> M.toast({html: 'Passwords did not match!', classes: 'red darken-4'}); </script>";
 			}
-	    }
+		}
 
+		if(isset($loginMessage)) { 
+			echo $loginMessage;
+		}
+		if(isset($registerError)) { 
+		 echo $registerError;
+		}
+		if(isset($registerError)) { 
+		 echo $registerError;
+		}
       ?>
    </body>
 </html>
